refactor: split header to small chunks
