+++ src/components/header/Header.tsx
@@ -1,24 +1,13 @@
-import LoginIcon from '@mui/icons-material/LoginRounded';
-import LogoutIcon from '@mui/icons-material/LogoutRounded';
-import ProfileIcon from '@mui/icons-material/ManageAccountsSharp';
 import MenuRoundedIcon from '@mui/icons-material/MenuRounded';
 import { Container, IconButton } from '@mui/material';
 import Box from '@mui/material/Box';
-import Button from '@mui/material/Button';
-import Stack from '@mui/material/Stack';
 import { styled } from '@mui/material/styles';
 import { useState } from 'react';
 import { NavLink } from 'react-router';
 
-import { useAppDispatch, useAppSelector } from '../../redux/hooks.ts';
-import { getCustomer } from '../../redux/selectors.ts';
-import { logOut } from '../../redux/slices/authSlice.ts';
-import CONSTANTS from '../../utils/CONSTANTS.ts';
 import { Navigation } from '../index.ts';
-import Cart from './Cart/Cart.tsx';
 import AsidePanel from './DropDownPanel.tsx';
 import LogoMain from './LogoMain';
-import Magnifier from './Magnifier/Magnifier.tsx';
 
 const StyledHeader = styled('header')(({ theme }) => ({
   backgroundColor: theme.palette.background.default,
@@ -31,20 +20,14 @@ const StyledHeader = styled('header')(({ theme }) => ({
 }));
 
 const Header = () => {
-  const { isLoading, customer } = useAppSelector(getCustomer);
   const [isOpen, setIsOpen] = useState(false);
-  const dispatch = useAppDispatch();
 
   return (
     <StyledHeader>
       <Container
         sx={{
-          // margin: '0 auto',
           gridTemplateColumns: { xs: '1fr', sm: '1fr auto 1fr' },
-          // gap: '1rem',
-          // alignItems: 'center',
           padding: '1vw',
-          // breakAfter:
           display: { xs: 'flex', sm: 'grid' },
           justifyContent: 'space-between',
         }}
@@ -73,54 +56,6 @@ const Header = () => {
             display: { xs: 'none', md: 'flex' },
           }}
         />
-        <Stack
-          direction="row"
-          sx={{
-            display: { xs: 'none', sm: 'flex' },
-            alignItems: 'center',
-            justifyContent: { xs: 'center', sm: 'flex-end' },
-            gap: 'clamp(0.5rem, 2.2vw, 2rem)',
-          }}
-        >
-          <IconButton>
-            <Magnifier />
-          </IconButton>
-          <IconButton>
-            <Cart />
-          </IconButton>
-          <IconButton component={NavLink} to={CONSTANTS.profile}>
-            <ProfileIcon color={customer ? 'primary' : 'action'} />
-          </IconButton>
-
-          {customer ? (
-            <Button
-              component={NavLink}
-              endIcon={<LogoutIcon />}
-              onClick={() => dispatch(logOut())}
-              sx={{
-                display: { xs: 'none', sm: 'flex' },
-                fontSize: { xs: 0, md: '1rem' },
-              }}
-              to="/login"
-              variant="contained"
-            >
-              Logout
-            </Button>
-          ) : (
-            <Button
-              component={NavLink}
-              startIcon={<LoginIcon />}
-              sx={{
-                display: { xs: 'none', sm: 'flex' },
-                fontSize: { xs: 0, md: '1rem' },
-              }}
-              to="/register"
-              variant="contained"
-            >
-              {isLoading ? 'loading' : 'Login/Register'}
-            </Button>
-          )}
-        </Stack>
         {isOpen ? <AsidePanel isOpen={isOpen} setIsOpen={setIsOpen} /> : null}
       </Container>
     </StyledHeader>
